bento v0.9.2 Bento

An incredibly fast, correct, pure-Elixir Bencoding library.

This module contains high-level methods to encode and decode Bencoded data.

Functions

decode(iodata, options \\ [])

Specs

decode(iodata, Keyword.t) ::
  {:ok, Bento.Parser.t} |
  {:error, :invalid} |
  {:error, {:invalid, String.t}}

Decode bencoded data to a value.

iex> Bento.decode("li1e3:twoli3eee")
{:ok, [1, "two", [3]]}

decode!(iodata, options \\ [])

Specs

decode!(iodata, Keyword.t) ::
  Bento.Parser.t |
  no_return

Decode bencoded data to a value, raising an exception on error.

iex> Bento.decode!("li1e3:twoli3eee")
[1, "two", [3]]

encode(value, options \\ [])

Specs

encode(Bento.Encoder.t, Keyword.t) ::
  {:ok, iodata} |
  {:ok, String.t} |
  {:error, {:invalid, any}}

Bencode a value.

iex> Bento.encode([1, "two", [3]])
{:ok, "li1e3:twoli3eee"}

encode!(value, options \\ [])

Specs

encode!(Bento.Encoder.t, Keyword.t) ::
  iodata |
  String.t |
  no_return

Bencode a value, raising an exception on error.

iex> Bento.encode!([1, "two", [3]])
"li1e3:twoli3eee"

encode_to_iodata(value, options \\ [])

Specs

encode_to_iodata(Bento.Encoder.t, Keyword.t) ::
  {:ok, iodata} |
  {:error, {:invalid, any}}

Bencode a value as iodata.

iex> Bento.encode_to_iodata([1, "two", [3]])
{:ok, [108, [[105, "1", 101], ["3", 58, "two"], [108, [[105, "3", 101]], 101]], 101]}

encode_to_iodata!(value, options \\ [])

Specs

encode_to_iodata!(Bento.Encoder.t, Keyword.t) ::
  iodata |
  no_return

Bencode a value as iodata, raises an exception on error.

iex> Bento.encode_to_iodata!([1, "two", [3]])
[108, [[105, "1", 101], ["3", 58, "two"], [108, [[105, "3", 101]], 101]], 101]

torrent(iodata)

Like decode, but ensures the data is a valid torrent metainfo file.

torrent!(iodata)

Like decode!, but ensures the data is a valid torrent metainfo file.